requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
reportlab==4.0.4
scikit-learn==1.3.0
pandas==2.0.3
bcrypt==4.0.1
mysql-connector-python==8.2.0
